Fix paho-mqtt dependency, rename config files to be app specific, rename app and service to use hyphens
