+++ .fpm
@@ -5,11 +5,13 @@
 --log warn
 --python-bin python3
 --python-package-name-prefix python3
+--python-disable-dependency paho-mqtt
 --python-disable-dependency PyYAML
---depends python3-yaml
 --python-disable-dependency pytz
+--depends python3-yaml
 --depends python3-tz
 --depends libcap-dev
---deb-systemd service/sentinel_mrhat_cam.service
+--before-install script/sentinel-mrhat-cam.preinst
+--deb-systemd service/sentinel-mrhat-cam.service
 --deb-systemd-enable
 --deb-systemd-auto-start
