+++ bin/sentinel_mrhat_cam.sh
