+++ script/sentinel-mrhat-cam.preinst
@@ -1,0 +1,1 @@
+pip3 install --upgrade paho-mqtt
